Refactor CPU struct with named fields and remove serialization
Replace the opaque byte-array CPU representation with named register
fields, removing the manual `as_bytes`/`from_bytes` serialization
methods. Simplify `update_zn_flags` logic and fix APU tick type
mismatch.

+++ src/lib.rs
@@ -60,7 +60,7 @@ pub fn tick(cpu: &mut CpuRp2a03, bus: &mut Bus<'_>) -> u8 {
     let cycles = TABLE[opcode as usize](cpu, bus);
 
     bus.ppu.tick_batch((cycles as u16) * 3);
-    bus.apu.tick(cycles);
+    bus.apu.tick(cycles as u16);
 
     if bus.poll_nmi() {
         nmi(cpu, bus);
